função

+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: mgomes-s <mgomes-s@student.42.rio>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/23 15:21:43 by mgomes-s          #+#    #+#             */
-/*   Updated: 2024/09/23 16:48:29 by mgomes-s         ###   ########.fr       */
+/*   Updated: 2024/10/07 15:26:59 by mgomes-s         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
-	unsigned int		i;
+	size_t				i;
 	char				*ptr;
 	const char			*ptr2;
